Backend_Service/requirements.txt
Flask
SQLAlchemy
flask_sqlalchemy
psycopg2-binary
